Added plain keyboard mechanism for inline buttons and reply buttons, changed keyboard id type in message from long to int

+++ src/main/java/kz/rbots/bekertugan/telegrambot/entity/fundamental/Message.java
@@ -15,7 +15,7 @@ public class Message {
     @Id
     private long   id;
     private String messageText;
-    private long   keyboardId;
+    private int    keyboardId;
 
     public Message() {
     }
